fix: address space cleanup and closing of open files may result in disk operations and cause threads to attempt to acquire same lock twice
Address space cleanup may result in SWAP files being removed, and
closing open files may result in their deletion if they have previously
been marked for deletion and the thread was the last with an open
reference to it. If the thread performing the cleanup invoked SWITCH
after acquiring the lock on SynchDisk, attempting to perform any of
these disk operations would cause the thread to attempt to acquire the
lock again, which is prevented by an assertion in the lock
implementation to avoid deadlocks.